backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
firebase-admin==6.7.0
httpx==0.28.1
python-multipart==0.0.20
pypdf==5.9.0
tavily-python==0.7.12
duckduckgo-search==8.1.1
langchain==0.3.27
langchain-openai==0.3.30
langchain-community==0.3.29
resend==2.13.0
sendgrid==6.12.4
